# openapi.yaml
openapi: 3.0.3
info:
  title: PartyHub API
  description: Social event platform for organizing and discovering parties, meetups, and social gatherings
  version: 1.0.0
  contact:
    name: PartyHub Team
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /api/users:
    get:
      summary: List or search users
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query (searches distinctName)
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDto'
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/{id}:
    get:
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDto'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/handle/{distinctName}:
    get:
      summary: Get user by handle
      parameters:
        - name: distinctName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/username/{username}:
    get:
      summary: Get user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/me:
    get:
      summary: Get current user
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/location/{id}:
    get:
      summary: Get user location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: User ID
      responses:
        '200':
          description: User location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLocation'
        '404':
          description: Location not found
  /api/users/location:
    put:
      summary: Update current user location
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdateDto'
      responses:
        '200':
          description: Location updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLocation'
        '400':
          description: User ID required
        '404':
          description: User not found
  /api/users/{id}/media:
    get:
      summary: Get user media
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of user media
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Media'
  /api/users/{id}/followers:
    get:
      summary: Get user followers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of followers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /api/users/{id}/following:
    get:
      summary: Get users that user is following
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of following
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /api/users/{id}/followers/count:
    get:
      summary: Get followers count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Followers count
  /api/users/{id}/following/count:
    get:
      summary: Get following count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Following count
  /api/users/{id}/follow-requests:
    get:
      summary: Get pending follow requests
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pending requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /api/users/{userId1}/followers/{userId2}/status:
    get:
      summary: Check follow status
      parameters:
        - name: userId1
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId2
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Follow status
          content:
            application/json:
              schema:
                type: boolean
  /api/users/{id}/follow:
    post:
      summary: Send follow request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: targetUserId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Request sent
  /api/users/{id}/followers/{followerId}:
    put:
      summary: Accept follow request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: followerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Request accepted
    delete:
      summary: Remove follow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: followerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Follow removed
  /api/users/{id}/profile-picture:
    get:
      summary: Get user profile picture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Profile picture
          content:
            image/*:
              schema:
                type: string
                format: binary
  /api/users/{id}/profile-picture-filename:
    get:
      summary: Get profile picture filename
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Filename
  /api/users/{id}/upload-profile-picture:
    post:
      summary: Upload profile picture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Picture uploaded
...
  /api/qr/image/user/{userId}:
    get:
      summary: Get QR image by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: QR image
          content:
            image/png:
              schema:
                type: string
                format: binary
...
components:
  schemas:
    Party:
      type: object
      properties:
        id:
          type: integer
          format: int64
        host_user:
          $ref: '#/components/schemas/User'
        category:
          $ref: '#/components/schemas/Category'
        title:
          type: string
        time_start:
          type: string
          format: date-time
        time_end:
          type: string
          format: date-time
        max_people:
          type: integer
        min_age:
          type: integer
        max_age:
          type: integer
        website:
          type: string
        description:
          type: string
        fee:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/Location'
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        displayName:
          type: string
        distinctName:
          type: string
        email:
          type: string
        biography:
          type: string
    UserCreateDto:
      type: object
      properties:
        displayName:
          type: string
        distinctName:
          type: string
        email:
          type: string
        biography:
          type: string
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
    Media:
      type: object
      properties:
        id:
          type: integer
          format: int64
        file:
          type: string
        party:
          $ref: '#/components/schemas/Party'
...
    UserLocation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    LocationUpdateDto:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
...